Initial Synergies propagation
